hw: Widen the irqs field of AcpiGetResourcesReply to uint32
The field needs to be able to hold GSIs.

+++ protocols/hw/include/protocols/hw/client.hpp
@@ -74,7 +74,7 @@ struct BatteryState {
 
 struct AcpiResources {
 	std::vector<uint16_t> io_ports;
-	std::vector<uint8_t> irqs;
+	std::vector<uint32_t> irqs;
 };
 
 struct DtRegister {
